Refactor: clean up code (service, controller layers)

+++ src/main/java/com/example/practice/dto/product/CreateProductRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record CreateProductRequest(
         @NotNull(message = "Поставщик обязателен")
         Long supplierId,
@@ -12,6 +14,6 @@ public record CreateProductRequest(
         @NotBlank
         String description,
         @Positive(message = "Цена не может быть отрицательной")
-        Double price
+        BigDecimal price
 ) {
 }
